Checkstyle on package names can sometimes be a pain...
