+++ buildSrc/build.gradle
@@ -1,9 +1,19 @@
 plugins {
 	id "java"
+	id "checkstyle"
 	id "org.cadixdev.licenser" version "0.5.0"
+}
+
+repositories {
+	mavenCentral()
 }
 
 license {
 	header project.file("../HEADER")
 	include "**/*.java"
+}
+
+checkstyle {
+	configFile = rootProject.file("../checkstyle.xml")
+	toolVersion = "8.31"
 }
